[FIX] base,http_routing: move keep_query method
It is used by some modules like auth_signup that do not depend of
http_routing
It was moved from base to http_routing at 880954ebfc110 but it should
have stayed in base

Before this commit it was not possible to access the /web/login page
when only the auth_signup module is installed.
