+++ odoo/addons/base/models/ir_qweb.py
@@ -367,6 +367,7 @@ structure.
 
 """
 
+import fnmatch
 import logging
 import re
 import traceback
@@ -442,6 +443,26 @@ SPECIAL_DIRECTIVES = {'t-translation', 't-ignore', 't-title'}
 # Name of the variable to insert the content in t-call in the template.
 # The slot will be replaced by the `t-call` tag content of the caller.
 T_CALL_SLOT = '0'
+
+
+def keep_query(*keep_params, **additional_params):
+    """
+    Generate a query string keeping the current request querystring's parameters specified
+    in ``keep_params`` and also adds the parameters specified in ``additional_params``.
+    Multiple values query string params will be merged into a single one with comma seperated
+    values.
+    The ``keep_params`` arguments can use wildcards too, eg:
+        keep_query('search', 'shop_*', page=4)
+    """
+    if not keep_params and not additional_params:
+        keep_params = ('*',)
+    params = additional_params.copy()
+    qs_keys = list(request.httprequest.args) if request else []
+    for keep_param in keep_params:
+        for param in fnmatch.filter(qs_keys, keep_param):
+            if param not in additional_params and param in qs_keys:
+                params[param] = request.httprequest.args.getlist(param)
+    return werkzeug.urls.url_encode(params)
 
 
 def indent_code(code, level):
@@ -793,6 +814,7 @@ class IrQWeb(models.AbstractModel):
             values.setdefault('res_company', self.env.company.sudo())
 
             values.update(
+                keep_query=keep_query,
                 request=request,  # might be unbound if we're not in an httprequest context
                 test_mode_enabled=bool(config['test_enable'] or config['test_file']),
                 json=scriptsafe,
